+++ src/redux/actions/index.js
@@ -1,10 +1,21 @@
+import * as accountsActions from './accountsActions';
 import * as appActions from './appActions';
+import * as assetsActions from './assetsActions';
 import * as blocksActions from './blocksActions';
+import * as faucetActions from './faucetActions';
+import * as governanceActions from './governanceActions';
 import * as txsActions from './txsActions';
 import * as validatorsActions from './validatorsActions';
-import * as accountsActions from './accountsActions';
-import * as assetsActions from './assetsActions';
-import * as faucetActions from './faucetActions';
 import * as xhrActions from './xhrActions';
 
-export { appActions, xhrActions, blocksActions, txsActions, validatorsActions, accountsActions, assetsActions, faucetActions };
+export {
+  accountsActions,
+  appActions,
+  assetsActions,
+  blocksActions,
+  faucetActions,
+  governanceActions,
+  txsActions,
+  validatorsActions,
+  xhrActions,
+};